requirements.txt
streamlit==1.46.1
openai==1.93.0
python-dotenv==1.1.1
